requirements.txt
pymavlink>=2.4.37
PySide6>=6.5.0
PySide6-WebEngine>=6.5.0
pyserial>=3.5
pytest>=7.4.0
pytest-qt>=4.2.0
folium>=0.14.0 